minor feedback
